hookUpdate: more context updates, trying to figure out where to put some logic
