+++ src/components/GameOptions.tsx
@@ -2,13 +2,8 @@ import React, { useContext } from "react";
 import { Button, InputNumber } from "antd";
 import { GameContext } from "./GameContext";
 
-export interface GameOptionsProps {
-  createBoard: () => void;
-}
-
-const GameOptions = (props: GameOptionsProps) => {
-  const { rows, cols, setCols, setRows } = useContext(GameContext);
-  const { createBoard } = props;
+const GameOptions = () => {
+  const { rows, cols, setCols, setRows, createBoard } = useContext(GameContext);
 
   return (
     <div>
@@ -24,6 +19,7 @@ const GameOptions = (props: GameOptionsProps) => {
         addonBefore="Column(s)"
         size="small"
         value={cols}
+        defaultValue={cols}
         onChange={setCols}
       />
       <Button onClick={() => createBoard()}> Create Board </Button>
